requirements.txt
tensorflow-gpu==1.12.0
keras==2.2.4
numpy
tqdm
opencv-python
comet_ml
scikit-learn
pillow
editdistance
dotmap
textdistance
